FEM: example multimaterial, move mesh from unit tests to fem examples
