+++ src/Mod/Fem/femexamples/meshes/mesh_multimaterial_twoboxes.py
